Fixing size for g_file_get_contents
Indeed in gphoto2-file.h for `gp_file_get_data_and_size` the size is defined as
unsigned long int. This is also fine for g_file_get_contents

+++ src/common/camera_control.c
@@ -1030,7 +1030,7 @@ void dt_camctl_import(const dt_camctl_t *c, const dt_camera_t *cam, GList *image
     CameraFile* camfile;
     int res = GP_OK;
     char *data = NULL;
-    size_t size;
+    unsigned long int size;
     time_t exif_time;
     if(!sdcard)
     {
